tsne-package {tsne}	R Documentation

The tsne-package for multidimensional scaling

Description

This package contains one function called tsne which contains all the functionality.

Details

Package: 	 tsne

Type: 	 Package

Version: 	 0.1

Date: 	 2010-02-19

License: 	 GPL

LazyLoad: 	 yes

Author(s)

Justin Donaldson https://github.com/jdonaldson/rtsne Maintainer: Justin Donaldson (jdonaldson@gmail.com)

References

L.J.P. van der Maaten and G.E. Hinton. Visualizing High-Dimensional Data Using t-SNE. Journal of Machine Learning Research 9 (Nov) : 2579-2605, 2008.

L.J.P. van der Maaten. Learning a Parametric Embedding by Preserving Local Structure. In Proceedings of the Twelfth International Conference on Artificial Intelligence and Statistics (AISTATS), JMLR W&CP 5:384-391, 2009.

[Package tsne version 0.1-3.1 Index]